requirements.txt
Django===4.0.6
djangorestframework==3.13.1
mysqlclient==2.1.1
django-mysql==4.7.0
django-cors-headers==3.13.0
pika==1.3.0
pytest-django==4.5.2